jupyter/notebooks/simulate_pipeline/pyproject.toml
[tool.poetry]
name = "simulate_pipeline"
version = "0.1.0"
description = ""
authors = ["cltj <tj@cloudlink.ai>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8,<3.11"
python-dotenv = "^0.20.0"
pandas = "^1.4.3"
numpy = "^1.23.2"
matplotlib = "^3.5.3"
seaborn = "^0.11.2"
python-pptx = "^0.6.21"
plotly = "^5.10.0"
pytest = "^7.1.2"
google-cloud-bigquery = "^3.3.2"
pandas-gbq = "^0.17.8"
azure-keyvault-secrets = "^4.5.1"
pydantic = "^1.9.2"
mkdocs = "^1.3.1"
azure-core = "^1.25.0"
tqdm = "^4.64.1"
ipykernel = "^6.15.3"
pandas-profiling = "^3.3.0"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"
azure-identity = "^1.11.0"
azure-mgmt-consumption = "^10.0.0"
boto3 = "^1.24.82"
azure-storage-blob = "^12.13.1"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.20.0"
pandas = "^1.4.3"
numpy = "^1.23.2"
matplotlib = "^3.5.3"
seaborn = "^0.11.2"
python-pptx = "^0.6.21"
plotly = "^5.10.0"
pytest = "^7.1.2"
google-cloud-bigquery = "^3.3.2"
azure-identity = "^1.10.0"
azure-mgmt-consumption = "^10.0.0"
azure-storage-blob = "^12.13.1"
boto3 = "^1.24.62"
pandas-gbq = "^0.17.8"
azure-keyvault-secrets = "^4.5.1"
pydantic = "^1.9.2"
mkdocs = "^1.3.1"
ipykernel = "^6.15.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
